feat(system): Move GetPrettyName function to system package
for macOS, Linux, and Windows
