+++ internal/system/host_linux.go
@@ -1,7 +1,7 @@
 //go:build linux
 // +build linux
 
-package metric
+package system
 
 import (
 	"bufio"
